+++ priceserver-ts/src/lib/error.ts
@@ -1,7 +1,7 @@
 import * as sentry from '@sentry/node';
 import * as logger from './logger';
 
-export function errorHandling(error: any) {
+export function errorHandler(error: any) {
   sentry.captureException(error);
   logger.error(error);
 }
