+++ priceserver-ts/src/lib/statistics.ts
@@ -1,4 +1,27 @@
-import { vwap } from './vwap';
+export function average(array: number[]): number {
+  if (!array || !array.length) {
+    throw new Error('empty array');
+  }
+
+  if (array.length === 1) {
+    return array[0];
+  }
+
+  return array.reduce((a, b) => a + b) / array.length;
+}
+
+export function vwap(array: { price: number; volume: number }[]): number {
+  if (!array || !array.length) {
+    throw new Error('empty array');
+  }
+
+  if (array.length === 1) {
+    return array[0].price;
+  }
+
+  // sum(volume * price) / (total volume)
+  return array.reduce((s, x) => s + x.volume * x.price, 0) / array.reduce((s, x) => s + x.volume, 0) || 0;
+}
 
 export function tvwap(
   array: { price: number; volume: number; timestamp: number }[],
